dht+route: add prel. ping on full buckets
